Refactored and removed redundant Cycle Screen code
