+++ app/src/main/java/com/tpp/theperiodpurse/ui/cycle/PeriodHistory.kt
@@ -16,22 +16,16 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.semantics.contentDescription
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 import com.tpp.theperiodpurse.ui.viewmodel.AppViewModel
 import com.tpp.theperiodpurse.R
-import com.tpp.theperiodpurse.data.addOneDay
-import com.tpp.theperiodpurse.data.entity.Date
 import com.tpp.theperiodpurse.data.findYears
 import com.tpp.theperiodpurse.data.parseDatesIntoPeriods
+import com.tpp.theperiodpurse.ui.cycle.components.PeriodEntries
+import com.tpp.theperiodpurse.ui.cycle.components.YearTab
 import com.tpp.theperiodpurse.ui.onboarding.scaledSp
-import java.text.SimpleDateFormat
 import java.time.LocalDate
-import java.time.ZoneId
-
-import java.util.*
 import kotlin.collections.ArrayList
 
 @Composable
@@ -49,7 +43,7 @@ fun PeriodHistoryLayout(
         years[LocalDate.now().year] = ArrayList()
     }
     var yearSelected by remember { mutableStateOf(years.keys.last()) }
-    // iterateing over years and create horizontally scrollable buttons
+    // Iterateing over years and create horizontally scrollable buttons
     Scaffold(
         topBar = {
             TopAppBar (
@@ -74,7 +68,7 @@ fun PeriodHistoryLayout(
             )
         },
     )
-    { it ->
+    {
         Box(
             modifier = Modifier.padding(it)
         ) {
@@ -94,7 +88,7 @@ fun PeriodHistoryLayout(
                     .padding(horizontal = 20.dp, vertical = 25.dp)
             ) {
                 LazyRow {
-                    // reverse keys
+                    // Reverse keys
                     val reversedKeys = years.keys.reversed()
                     for (key in reversedKeys) {
                         item {
@@ -114,7 +108,7 @@ fun PeriodHistoryLayout(
                             horizontalArrangement = Arrangement.SpaceBetween
                         ) {
                             Text(
-                                // replace with year
+                                // Replace the year
                                 text = yearSelected.toString(),
                                 fontSize = 15.scaledSp(),
                                 fontWeight = FontWeight(700),
@@ -127,83 +121,11 @@ fun PeriodHistoryLayout(
                                 .padding(top = 5.dp, bottom = 10.dp)
                                 .fillMaxWidth()
                         )
-                        // show last three most recent periods
+                        // Show last three most recent periods
                         years[yearSelected]?.let { it1 -> PeriodEntries(it1, null) }
                     }
                 }
             }
         }
     }
-}
-
-@Composable
-fun PeriodEntries(periods: ArrayList<ArrayList<Date>>, upperbound: Int?) {
-    if (periods.size == 0 || periods[0].size == 0) {
-        Text(text = stringResource(R.string.please_start_logging_to_learn_more))
-    } else {
-        val length = if (upperbound != null) kotlin.math.min(periods.size, upperbound) else periods.size
-        periods.reverse()
-        Column {
-            val formatter = SimpleDateFormat("MMM d", Locale.getDefault())
-            // check if it's within a day, if so display current period...
-            for (i in 0 until length) {
-                val date = periods[i][0].date
-                val converted = date?.toInstant()?.atZone(ZoneId.systemDefault())?.toLocalDate()?.year
-                val current = LocalDate.now().year
-                if (i == 0 && converted == current) {
-                    val formattedDate = formatter.format(date)
-                    // only if same year
-                    Text(text = "Most Recent Period: Started $formattedDate")
-                } else {
-                    val period = periods[i]
-                    val startDate = period[0].date
-                    val endDate =
-                        period[period.size - 1].date?.let { addOneDay(it) }
-                    if (startDate != null && endDate != null) {
-                        val startString = formatter.format(startDate)
-                        val endString = formatter.format(endDate)
-                        val periodLength =
-                            (endDate.time - startDate.time) / 86400000
-                        Text(text = "$startString - $endString")
-                        Text(
-                            text = "$periodLength-day period",
-                            fontSize = 13.scaledSp()
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun YearTab(year: Int, selected: Boolean = false, onClick: () -> Unit) {
-    val red = Color(0xFFB31F20)
-    val outlineColor = if (selected) red else Color.LightGray
-    val contentColor = if (selected) Color.White else Color.LightGray
-    val backgroundColor = if (selected) red else Color.White
-    Button(
-        onClick = onClick,
-        border = BorderStroke(1.dp, outlineColor),
-        shape = RoundedCornerShape(25),
-        colors = ButtonDefaults.buttonColors(
-            backgroundColor = backgroundColor,
-            contentColor = contentColor
-        ),
-        modifier = Modifier.padding(start = 5.dp, end = 5.dp, bottom = 10.dp)
-    ){
-        Text( text = year.toString() )
-    }
-}
-
-@Preview
-@Composable
-fun YearTabFalsePreview() {
-    YearTab(year = 2023, selected = false) {}
-}
-
-@Preview
-@Composable
-fun YearTabTruePreview() {
-    YearTab(year = 2023, selected = true) {}
 }
